No-sort aggregation logic
